insightbot/requirements.txt
streamlit
yfinance
pandas
numpy
matplotlib
seaborn
google-cloud-bigquery
google-cloud-storage
adk